# 修正woc openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "修正woc",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/login": {
      "post": {
        "summary": "登录（必做）",
        "deprecated": false,
        "description": "##### 用户登录接口\n- 使用jwt令牌\n- 按照Bearer Token规范使用",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "userCode",
                  "password"
                ],
                "x-apifox-orders": [
                  "userCode",
                  "password"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "token"
                      ],
                      "x-apifox-orders": [
                        "token"
                      ],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "success": true,
                      "errCode": 0,
                      "errMsg": "string",
                      "data": {
                        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                      }
                    }
                  }
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [],
        "x-apifox-folder": "修正woc",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845400-run"
      }
    },
...
    "/register": {
      "post": {
        "summary": "注册(新)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "userCode",
                  "password"
                ],
                "required": [
                  "userCode",
                  "password"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "x-apifox-orders": [
                    "01JMVC5R9X9BEPWGQREJCSF1R0"
                  ],
                  "x-apifox-refs": {
                    "01JMVC5R9X9BEPWGQREJCSF1R0": {
                      "$ref": "#/components/schemas/GlobalResponse"
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ]
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc",
        "x-apifox-status": "developing",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263849724-run"
      }
    },
    "/user/userdetail": {
      "post": {
        "summary": "添加或更新个人信息(新)",
        "deprecated": false,
        "description": "更改的个人信息为携带token者的",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "studentId": {
                    "type": "string"
                  },
                  "academyId": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "x-apifox-orders": [
                  "name",
                  "studentId",
                  "academyId",
                  "phone"
                ],
                "required": [
                  "name",
                  "studentId",
                  "academyId",
                  "phone"
                ],
                "x-apifox-ignore-properties": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc",
        "x-apifox-status": "developing",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263857678-run"
      }
    }
  },
  "components": {
    "schemas": {
...
      "GlobalResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errCode": {
            "type": "integer"
          },
          "errMsg": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {},
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "success",
          "data",
          "errMsg",
          "errCode"
        ],
        "x-apifox-orders": [
          "success",
          "errCode",
          "errMsg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}